backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

